// subject_backend/src/subjects/enum/sort.enum.ts
export enum SubjectOrderBy {
  NOMBRE_ASC = 'nombre_asc',
  NOMBRE_DESC = 'nombre_desc',
  MATRICULA_ASC = 'matricula_asc',
  MATRICULA_DESC = 'matricula_desc',
  INICIO_MATRICULA_ASC = 'inicio_matricula_asc',
  INICIO_MATRICULA_DESC = 'inicio_matricula_desc',
  FIN_MATRICULA_ASC = 'fin_matricula_asc',
  FIN_MATRICULA_DESC = 'fin_matricula_desc',
  INICIO_ASIGNATURA_ASC = 'inicio_asignatura_asc',
  INICIO_ASIGNATURA_DESC = 'inicio_asignatura_desc',
  FIN_ASIGNATURA_ASC = 'fin_asignatura_asc',
  FIN_ASIGNATURA_DESC = 'fin_asignatura_desc'
}